@@ -1,13 +1,13 @@
 {
   "name": "@lab900/ui",
-  "version": "16.1.28",
+  "version": "17.0.0",
   "author": "Lab900 <info@lab900.com> (https://lab900.com)",
   "licens": "MIT",
   "peerDependencies": {
-    "@angular/common": "^16.2.8",
-    "@angular/core": "^16.2.8",
-    "@angular/material": "^16.2.7",
-    "@ngx-translate/core": "^15.0.0"
+    "@angular/common": ">=17.0.0",
+    "@angular/core": ">=17.0.0",
+    "@angular/material": ">=17.0.0",
+    "@ngx-translate/core": ">=15.0.0"
   },
   "repository": {
     "type": "git",
