@@ -1,6 +1,6 @@
 {
   "name": "angular-library-ui",
-  "version": "16.0.0",
+  "version": "17.0.0",
   "repository": "https://github.com/lab900/angular-library-ui",
   "homepage": "https://lab900.github.io/angular-libraries",
   "author": "Lab900 <info@lab900.com> (https://lab900.com)",
@@ -21,58 +21,42 @@
     "pretty:html:fix": "prettier --no-config --print-width=140 --write \"{src,lib}/**/**.html\"",
     "clean": "rm -rf node_modules/ && npm install"
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged",
-      "pre-push": "ng lint --fix"
-    }
-  },
-  "lint-staged": {
-    "**/*.{js,ts,json,css,scss,html,md}": [
-      "prettier --write",
-      "git add"
-    ]
-  },
   "dependencies": {
-    "@angular/animations": "^16.2.8",
-    "@angular/cdk": "^16.2.7",
-    "@angular/common": "^16.2.8",
-    "@angular/compiler": "^16.2.8",
-    "@angular/core": "^16.2.8",
-    "@angular/forms": "^16.2.8",
-    "@angular/material": "^16.2.7",
-    "@angular/platform-browser": "^16.2.8",
-    "@angular/platform-browser-dynamic": "^16.2.8",
-    "@angular/router": "^16.2.8",
-    "@ngx-translate/core": "^14.0.0",
-    "moment": "^2.29.4",
-    "ngx-markdown": "^16.0.0",
+    "@angular/animations": "^17.3.5",
+    "@angular/cdk": "^17.3.5",
+    "@angular/common": "^17.3.5",
+    "@angular/compiler": "^17.3.5",
+    "@angular/core": "^17.3.5",
+    "@angular/forms": "^17.3.5",
+    "@angular/material": "^17.3.5",
+    "@angular/platform-browser": "^17.3.5",
+    "@angular/platform-browser-dynamic": "^17.3.5",
+    "@angular/router": "^17.3.5",
+    "@ngx-translate/core": "^15.0.0",
+    "ngx-markdown": "^17.2.0",
     "prismjs": "^1.29.0",
-    "rxjs": "~7.8.0",
+    "rxjs": "~7.8.1",
     "tslib": "^2.0.0",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.5",
-    "@angular-eslint/builder": "16.2.0",
-    "@angular-eslint/eslint-plugin": "16.2.0",
-    "@angular-eslint/eslint-plugin-template": "16.2.0",
-    "@angular-eslint/schematics": "16.2.0",
-    "@angular-eslint/template-parser": "16.2.0",
-    "@angular/cli": "^16.2.5",
-    "@angular/compiler-cli": "^16.2.8",
-    "@angular/language-service": "^16.2.8",
+    "@angular-devkit/build-angular": "^17.3.5",
+    "@angular-eslint/builder": "17.3.0",
+    "@angular-eslint/eslint-plugin": "17.3.0",
+    "@angular-eslint/eslint-plugin-template": "17.3.0",
+    "@angular-eslint/schematics": "17.3.0",
+    "@angular-eslint/template-parser": "17.3.0",
+    "@angular/cli": "^17.3.5",
+    "@angular/compiler-cli": "^17.3.5",
+    "@angular/language-service": "^17.3.5",
     "@types/jasmine": "~3.10.2",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "^18.15.5",
-    "@typescript-eslint/eslint-plugin": "^5.59.2",
-    "@typescript-eslint/parser": "^5.59.2",
-    "angular-cli-ghpages": "^1.0.5",
-    "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^8.39.0",
+    "@types/node": "^20.11.1",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^3.4.0",
-    "husky": "^4.3.8",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.4.1",
@@ -80,11 +64,10 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "lint-staged": "^10.5.4",
-    "ng-packagr": "^16.2.3",
-    "prettier": "2.8.8",
+    "ng-packagr": "^17.3.0",
+    "prettier": "3.2.5",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.9.5"
+    "typescript": "~5.4.5"
   }
 }
