@@ -6,16 +6,19 @@ import {
 } from '@angular/core';
 import { Icon } from '../../models/nav-item.model';
 import { MatIconModule } from '@angular/material/icon';
-import { NgIf } from '@angular/common';
 
 @Component({
   selector: 'lab900-icon',
   styleUrls: ['./icon.component.scss'],
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  template: ` <mat-icon *ngIf="icon?.name">{{ icon.name }}</mat-icon>
-    <mat-icon *ngIf="icon?.svgName" [svgIcon]="icon.svgName"></mat-icon>`,
-  imports: [MatIconModule, NgIf],
+  template: ` @if (icon?.name) {
+      <mat-icon>{{ icon.name }}</mat-icon>
+    }
+    @if (icon?.svgName) {
+      <mat-icon [svgIcon]="icon.svgName"></mat-icon>
+    }`,
+  imports: [MatIconModule],
   standalone: true,
 })
 export class IconComponent {
