@@ -3,7 +3,7 @@ import {
   Component,
   ViewEncapsulation,
 } from '@angular/core';
-import { AsyncPipe, NgIf } from '@angular/common';
+import { AsyncPipe } from '@angular/common';
 import { CellRendererAbstract } from '../cell-renderer.abstract';
 import { CellWithAnchorRendererOptions } from './cell-with-anchor-renderer.options';
 import { combineLatest, Observable } from 'rxjs';
@@ -15,35 +15,30 @@ import { TranslateModule } from '@ngx-translate/core';
 @Component({
   selector: 'lab900-cell-with-anchor',
   standalone: true,
-  imports: [
-    NgIf,
-    AsyncPipe,
-    MatTooltipModule,
-    MatCheckboxModule,
-    TranslateModule,
-  ],
+  imports: [AsyncPipe, MatTooltipModule, MatCheckboxModule, TranslateModule],
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
-  template: `<div
-    class="lab900-cell-value lab900-cell-value--with-anchor"
-    *ngIf="href$ | async as href"
-    [matTooltip]="(tooltip$ | async) ?? ''"
-    [matTooltipPosition]="(tooltipPosition$ | async) ?? undefined"
-    (click)="$event.stopImmediatePropagation()"
-  >
-    <a [target]="(rendererOptions$ | async)?.target ?? '_self'" [href]="href">
-      {{ cellValue$ | async | translate }}
-    </a>
-  </div>`,
+  template: `@if (href$ | async; as href) {
+    <div
+      class="lab900-cell-value lab900-cell-value--with-anchor"
+      [matTooltip]="(tooltip$ | async) ?? ''"
+      [matTooltipPosition]="(tooltipPosition$ | async) ?? undefined"
+      (click)="$event.stopImmediatePropagation()"
+    >
+      <a [target]="(rendererOptions$ | async)?.target ?? '_self'" [href]="href">
+        {{ cellValue$ | async | translate }}
+      </a>
+    </div>
+  }`,
 })
 export class CellWithAnchorRendererComponent<
-  T = any
+  T = any,
 > extends CellRendererAbstract<CellWithAnchorRendererOptions<T>, T> {
   public readonly href$: Observable<string> = combineLatest([
     this.rendererOptions$,
     this.data$,
   ]).pipe(
     map(([options, data]) => options?.url(data) ?? ''),
-    shareReplay({ bufferSize: 1, refCount: true })
+    shareReplay({ bufferSize: 1, refCount: true }),
   );
 }
