@@ -17,17 +17,18 @@ import { MatTooltipModule } from '@angular/material/tooltip';
   imports: [CommonModule, MatCheckboxModule, MatTooltipModule],
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
-  template: `<mat-checkbox
-    *ngIf="rendererOptions$ | async as options"
-    [matTooltip]="tooltip$ | async"
-    [matTooltipPosition]="tooltipPosition$ | async"
-    [checked]="cellValue$ | async"
-    [disabled]="disabled$ | async"
-    [indeterminate]="options.indeterminate ?? false"
-    [color]="options.theme ?? 'primary'"
-    (change)="onValueChange($event.checked)"
-    (click)="$event.stopImmediatePropagation()"
-  ></mat-checkbox>`,
+  template: `@if (rendererOptions$ | async; as options) {
+    <mat-checkbox
+      [matTooltip]="tooltip$ | async"
+      [matTooltipPosition]="tooltipPosition$ | async"
+      [checked]="cellValue$ | async"
+      [disabled]="disabled$ | async"
+      [indeterminate]="options.indeterminate ?? false"
+      [color]="options.theme ?? 'primary'"
+      (change)="onValueChange($event.checked)"
+      (click)="$event.stopImmediatePropagation()"
+    ></mat-checkbox>
+  }`,
 })
 export class CheckboxCellRendererComponent extends CellRendererAbstract<CheckboxCellRendererOptions> {
   public readonly disabled$ = combineLatest([
@@ -35,8 +36,8 @@ export class CheckboxCellRendererComponent extends CellRendererAbstract<Checkbox
     this.data$,
   ]).pipe(
     map(
-      ([config, data]) => config?.cellEditorOptions?.disabled?.(data) ?? false
-    )
+      ([config, data]) => config?.cellEditorOptions?.disabled?.(data) ?? false,
+    ),
   );
   public onValueChange(newValue: boolean): void {
     this.columnConfig$
@@ -44,7 +45,7 @@ export class CheckboxCellRendererComponent extends CellRendererAbstract<Checkbox
       .subscribe(([config, data]) => {
         if (!this.handleValueChanged) {
           throw Error(
-            `No handleValueChanged method provided for column ${config.key}`
+            `No handleValueChanged method provided for column ${config.key}`,
           );
         }
         this.handleValueChanged?.(newValue, config, data);
