@@ -4,7 +4,7 @@ import {
   Input,
   ViewEncapsulation,
 } from '@angular/core';
-import { AsyncPipe, NgIf } from '@angular/common';
+import { AsyncPipe } from '@angular/common';
 import { Lab900TableService } from '../../services/table.service';
 import { combineLatest, Observable, ReplaySubject } from 'rxjs';
 import { distinctUntilChanged, filter, map, shareReplay } from 'rxjs/operators';
@@ -13,14 +13,16 @@ import { TableCell } from '../../models/table-cell.model';
 @Component({
   selector: 'lab900-column-header-sorting',
   standalone: true,
-  imports: [NgIf, AsyncPipe],
+  imports: [AsyncPipe],
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  template: ` <ng-container *ngIf="sortIcon$ | async as icon">
-    <span class="material-icons lab900-sort-arrow" *ngIf="icon">
-      {{ icon }}
-    </span>
-  </ng-container>`,
+  template: ` @if (sortIcon$ | async; as icon) {
+    @if (icon) {
+      <span class="material-icons lab900-sort-arrow">
+        {{ icon }}
+      </span>
+    }
+  }`,
   styles: [
     `
       .lab900-sort-arrow {
@@ -52,8 +54,8 @@ export class ColumnHeaderSortingComponent {
       filter(([, col]) => !!col?.sortable),
       map(
         ([sort, col]) =>
-          sort?.find((s) => s.id === (col.sortKey ?? col.key))?.direction ?? ''
-      )
+          sort?.find((s) => s.id === (col.sortKey ?? col.key))?.direction ?? '',
+      ),
     );
 
     this.sortIcon$ = sortDirection$.pipe(
@@ -65,7 +67,7 @@ export class ColumnHeaderSortingComponent {
           return 'south';
         }
         return '';
-      })
+      }),
     );
   }
 }
