@@ -8,10 +8,11 @@ export * from './lib/utils/utils';
 export * from './lib/utils/different.utils';
 
 // dialogs
-export * from './lib/dialog/dialog.module';
 export * from './lib/dialog/components/alert-dialog/alert-dialog.component';
 export * from './lib/dialog/components/confirmation-dialog/confirmation-dialog.component';
 export * from './lib/dialog/directives/confirmation-dialog.directive';
+export * from './lib/dialog/models/alertDialog';
+export * from './lib/dialog/models/confirmationDialog';
 
 // alerts
 export * from './lib/alert/components/alert/alert.component';
@@ -23,9 +24,9 @@ export * from './lib/nav-list/components/nav-list/nav-list.component';
 export * from './lib/nav-list/components/icon/icon.component';
 
 // data list
-export * from './lib/data-list/data-list.module';
 export * from './lib/data-list/models/data-list.model';
 export * from './lib/data-list/components/data-list/data-list.component';
+export * from './lib/data-list/components/data-list-item/data-list-item.component';
 export * from './lib/data-list/directives/data-list-empty.directive';
 export * from './lib/data-list/directives/data-list-item-info.directive';
 
@@ -39,16 +40,17 @@ export * from './lib/page-header/components/page-header-nav-item/page-header-nav
 export * from './lib/page-header/models/page-header-nav.model';
 
 // merger
-export * from './lib/merger/merger.module';
 export * from './lib/merger/components/merger/merger.component';
+export * from './lib/merger/components/merger-item/merger-item.component';
 export * from './lib/merger/models/merge-object.model';
 export * from './lib/merger/models/merge-config.model';
 export * from './lib/merger/abstracts/custom-component.abstract';
 
 // buttons
-export * from './lib/button/button.module';
 export * from './lib/button/components/button/button.component';
 export * from './lib/button/components/action-button/lab900-action-button.component';
+export * from './lib/button/components/action-button-menu/lab900-action-button-menu.component';
+export * from './lib/button/components/action-button-toggle/lab900-action-button-toggle.component';
 export * from './lib/button/models/action-button.model';
 export * from './lib/button/models/button.model';
 
