@@ -1,11 +1,6 @@
 import { Component, ViewChild } from '@angular/core';
 import { mergerDataExample } from './config/merger-data-example';
-import {
-  Lab900MergerComponent,
-  Lab900MergerModule,
-  MergeConfig,
-  MergeObject,
-} from '@lab900/ui';
+import { Lab900MergerComponent, MergeConfig, MergeObject } from '@lab900/ui';
 import { mergerSchemaExample } from './config/merger-schema-example';
 import { MergerDataExample } from './models/merger-data-example.model';
 import { MatButtonModule } from '@angular/material/button';
@@ -14,7 +9,7 @@ import { TranslateModule } from '@ngx-translate/core';
 @Component({
   selector: 'lab900-merger-example',
   standalone: true,
-  imports: [MatButtonModule, TranslateModule, Lab900MergerModule],
+  imports: [MatButtonModule, TranslateModule, Lab900MergerComponent],
   template: `
     <div class="merger-example-header">
       <button
